+++ geo-types/CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## Unreleased
+
+- Implement Spade HasPosition for Coord
+  - <https://github.com/georust/geo/pull/1490>
+
 ## 0.7.18 - 2025-12-01
 
 - Renamed features `use-rstar`, `use-rstar-0_8`, etc. to simply `rstar`, `rstar-0_8`, etc. (removing the `use-` prefix)